+++ test/tangle/ResourceCommitments.t.sol
@@ -225,7 +225,7 @@ contract ResourceCommitmentsTest is BaseTest {
         bytes32 resourcesHash = SignatureLib.hashResourceCommitments(details.resourceCommitments);
 
         bytes32 QUOTE_TYPEHASH = keccak256(
-            "QuoteDetails(uint64 blueprintId,uint64 ttlBlocks,uint256 totalCost,uint64 timestamp,uint64 expiry,uint8 confidentiality,AssetSecurityCommitment[] securityCommitments,ResourceCommitment[] resourceCommitments)AssetSecurityCommitment(Asset asset,uint16 exposureBps)Asset(uint8 kind,address token)ResourceCommitment(uint8 kind,uint64 count)"
+            "QuoteDetails(address requester,uint64 blueprintId,uint64 ttlBlocks,uint256 totalCost,uint64 timestamp,uint64 expiry,uint8 confidentiality,AssetSecurityCommitment[] securityCommitments,ResourceCommitment[] resourceCommitments)AssetSecurityCommitment(Asset asset,uint16 exposureBps)Asset(uint8 kind,address token)ResourceCommitment(uint8 kind,uint64 count)"
         );
 
         bytes32 domainSeparator = keccak256(
@@ -241,6 +241,7 @@ contract ResourceCommitmentsTest is BaseTest {
         bytes32 structHash = keccak256(
             abi.encode(
                 QUOTE_TYPEHASH,
+                details.requester,
                 details.blueprintId,
                 details.ttlBlocks,
                 details.totalCost,